ファイル名の修正とadminのmockデータの配置ファイル変更
