+++ samples/web-csr/dressca-frontend/admin/mock/data/catalog-brands.ts
@@ -1,0 +1,16 @@
+import type { GetCatalogBrandsResponse } from '@/generated/api-client';
+
+export const catalogBrands: GetCatalogBrandsResponse[] = [
+  {
+    id: 1,
+    name: '高級なブランド',
+  },
+  {
+    id: 2,
+    name: 'カジュアルなブランド',
+  },
+  {
+    id: 3,
+    name: 'ノーブランド',
+  },
+];
