+++ samples/web-csr/dressca-frontend/admin/mock/data/catalog-categories.ts
@@ -1,0 +1,16 @@
+import type { GetCatalogCategoriesResponse } from '@/generated/api-client';
+
+export const catalogCategories: GetCatalogCategoriesResponse[] = [
+  {
+    id: 1,
+    name: '服',
+  },
+  {
+    id: 2,
+    name: 'バッグ',
+  },
+  {
+    id: 3,
+    name: 'シューズ',
+  },
+];
