+++ samples/web-csr/dressca-frontend/admin/mock/data/catalog-items.ts
