+++ samples/web-csr/dressca-frontend/admin/mock/data/users.ts
@@ -1,0 +1,7 @@
+import type { GetLoginUserResponse } from '@/generated/api-client';
+import { Roles } from '@/shared/constants/roles';
+
+export const user: GetLoginUserResponse = {
+  userName: 'admin@example.com',
+  roles: [Roles.ADMIN],
+};
