+++ samples/web-csr/dressca-frontend/admin/mock/handlers/assets-handler.ts
