+++ samples/web-csr/dressca-frontend/admin/mock/handlers/catalog-brands-handler.ts
@@ -1,6 +1,6 @@
 import { HttpStatusCode } from 'axios';
 import { HttpResponse, http } from 'msw';
-import { catalogBrands } from '../data/catalogBrands';
+import { catalogBrands } from '../data/catalog-brands';
 
 export const catalogBrandsHandlers = [
   http.get('/api/catalog-brands', () => {
