+++ samples/web-csr/dressca-frontend/admin/mock/handlers/catalog-categories-handler.ts
@@ -1,6 +1,6 @@
 import { HttpStatusCode } from 'axios';
 import { HttpResponse, http } from 'msw';
-import { catalogCategories } from '../data/catalogCategories';
+import { catalogCategories } from '../data/catalog-categories';
 
 export const catalogCategoriesHandlers = [
   http.get('/api/catalog-categories', () => {
