--- samples/web-csr/dressca-frontend/admin/mock/handlers/catalogBrands.ts
@@ -1,24 +1,0 @@
-import type { GetCatalogBrandsResponse } from '@/generated/api-client';
-import { HttpStatusCode } from 'axios';
-import { HttpResponse, http } from 'msw';
-
-const catalogBrands: GetCatalogBrandsResponse[] = [
-  {
-    id: 1,
-    name: '高級なブランド',
-  },
-  {
-    id: 2,
-    name: 'カジュアルなブランド',
-  },
-  {
-    id: 3,
-    name: 'ノーブランド',
-  },
-];
-
-export const catalogBrandsHandlers = [
-  http.get('/api/catalog-brands', () => {
-    return HttpResponse.json(catalogBrands, { status: HttpStatusCode.Ok });
-  }),
-];
