+++ samples/web-csr/dressca-frontend/consumer/mock/data/basket-items.ts
