+++ samples/web-csr/dressca-frontend/consumer/mock/data/catalog-brands.ts
