+++ samples/web-csr/dressca-frontend/consumer/mock/data/catalog-items.ts
