+++ samples/web-csr/dressca-frontend/consumer/mock/handlers.ts
@@ -1,9 +1,9 @@
-import { assetsHandlers } from './handlers/assetsHandler';
-import { basketsHandlers } from './handlers/basketsHandler';
-import { catalogItemsHandlers } from './handlers/catalogItemsHandler';
-import { catalogBrandsHandlers } from './handlers/catalogBrandsHandler';
-import { catalogCategoriesHandlers } from './handlers/catalogCategoriesHandler';
-import { orderingHandlers } from './handlers/orderingHandler';
+import { assetsHandlers } from './handlers/assets-handler';
+import { basketsHandlers } from './handlers/baskets-handler';
+import { catalogItemsHandlers } from './handlers/catalog-items-handler';
+import { catalogBrandsHandlers } from './handlers/catalog-brands-handler';
+import { catalogCategoriesHandlers } from './handlers/catalog-categories-handler';
+import { orderingHandlers } from './handlers/ordering-handler';
 
 export const handlers = [
   ...assetsHandlers,
