+++ samples/web-csr/dressca-frontend/consumer/mock/handlers/assets-handler.ts
