+++ samples/web-csr/dressca-frontend/consumer/mock/handlers/catalog-brands-handler.ts
@@ -1,0 +1,9 @@
+import { HttpStatusCode } from 'axios';
+import { HttpResponse, http } from 'msw';
+import { catalogBrands } from '../data/catalog-brands';
+
+export const catalogBrandsHandlers = [
+  http.get('/api/catalog-brands', () => {
+    return HttpResponse.json(catalogBrands, { status: HttpStatusCode.Ok });
+  }),
+];
