+++ samples/web-csr/dressca-frontend/consumer/mock/handlers/catalog-categories-handler.ts
@@ -1,0 +1,9 @@
+import { HttpStatusCode } from 'axios';
+import { HttpResponse, http } from 'msw';
+import { catalogCategories } from '../data/catalog-categories';
+
+export const catalogCategoriesHandlers = [
+  http.get('/api/catalog-categories', () => {
+    return HttpResponse.json(catalogCategories, { status: HttpStatusCode.Ok });
+  }),
+];
