+++ samples/web-csr/dressca-frontend/consumer/mock/handlers/ordering-handler.ts
